Added Join and Split functions
There is a bug in split_tree function due to erronous search() function.
Search is not able to find existing cal in tree.
Bug snapshot pasted at end of the code.
TODO : need to fix it.

Insert should also be fixed, on the same lines, not to include
duplicate entries in tree, by detecting if an entry already
exist in tree.
This behavior is erratic. Sometimes it allows adding duplicate
entries, sometimes doesn't.